==> 3. AWS Forecast/Modelo 3/3. Creating Forecast Predictor.ipynb <==
<center>
  <h2>Trabajo Laboratorio de Implementación III - MCD Virtual Cohorte 2022</h2>
  <h3>Forecasting para una Empresa de Consumo Masivo</h3>
  <h4>Entrenamiento de predictor</h4>
</center>

# Importación librerias

# !pip install awswrangler
# !pip install boto3
import boto3
import awswrangler as wr
import pandas as pd
import numpy as np
import warnings
from datetime import datetime
from dateutil.relativedelta import relativedelta
warnings.filterwarnings("ignore")

# Definición fechas

date_start=datetime.strptime('2017-01-01', '%Y-%m-%d')
date_end=datetime.strptime('2019-12-01', '%Y-%m-%d')
forecast_length = 2
date_future=date_end+relativedelta(months=forecast_length)
years=range(2018, 2020)

print(f"Fecha inicio historia {date_start}")
print(f"Fecha fin historia {date_end}")
print(f"Fecha futura (related time series) {date_future}")

timestamp_format="yyyy-MM-dd"
dataset_frequency='M'

# Versión modelo
data_version = '3'
grouped_cols = ["product_id"]
grouped_cols_name = "product"

# Conexiones AWS

account_id=boto3.client('sts').get_caller_identity().get('Account')
session = boto3.Session() 
forecast = session.client(service_name='forecast')
region = forecast.meta.region_name
forecastquery = session.client(service_name='forecastquery')
role = f'arn:aws:iam::{account_id}:role/ForecastRole'

# Qué valores predecir: media o percentiles
quantiles=['mean']
# Horizonte de predicción
forecast_length = 2
dataset_group_arn=f'arn:aws:forecast:us-east-1:{account_id}:dataset-group/modelo_{data_version}_{grouped_cols_name}'
dataset_group_arn

# Verificar comunicación con Amazon Forecast
assert forecast.list_predictors()

predictor_name_automl = f"modelo_{data_version}_predictor"

# Definir los parámetros del predictor
predictor_params = {
    'PredictorName': predictor_name_automl,
    'ForecastHorizon': forecast_length,  # Horizonte de predicción en meses
    'ForecastTypes': quantiles,  # Percentiles de predicción
    'PerformAutoML': True,
    'OptimizationMetric': 'MAPE',
    'InputDataConfig': {
        'DatasetGroupArn': dataset_group_arn,
    },
    'FeaturizationConfig': {
        'ForecastFrequency': dataset_frequency,  # Frecuencia de predicción (M para mensual)
        'Featurizations': [
            {
                'AttributeName': 'target_value',
                'FeaturizationPipeline': [
                    {
                        'FeaturizationMethodName': 'filling',
                        'FeaturizationMethodParameters': {
                            'aggregation': 'sum',
                            'frontfill': 'none',
                            'middlefill': 'zero',
                            'backfill': 'zero'
                        }
                    }
                ]
            }
        ]
    }
}

# Crear el predictor
response = forecast.create_predictor(**predictor_params)

automl_predictor_arn_wape = response['PredictorArn']

print(f"Waiting for Predictor with ARN {automl_predictor_arn_wape} to become ACTIVE. \nCurrent Status:")
# forecast.describe_auto_predictor(PredictorArn=automl_predictor_arn_wape)['Status']
forecast.describe_predictor(PredictorArn=automl_predictor_arn_wape)['Status']
